Get wasm bundling working
