+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "packages/core",
     "packages/cli",
     "packages/core-macro",
+    "packages/extension",
     "packages/router",
     "packages/html",
     "packages/hooks",
@@ -74,6 +75,7 @@ futures-channel = "0.3.21"
 futures-util = { version = "0.3", default-features = false }
 rustc-hash = "1.1.0"
 wasm-bindgen = "0.2.79"
+html_parser = "0.7.0"
 
 # This is a "virtual package"
 # It is not meant to be published, but is used so "cargo run --example XYZ" works properly
