--- packages/cli/extension/server/.gitignore
@@ -1,3 +1,0 @@
-**
-!Readme.md
-!.gitignore
