+++ packages/extension/static/icon.png
