+++ packages/extension/tsconfig.json
@@ -2,11 +2,17 @@
 	"compilerOptions": {
 		"module": "commonjs",
 		"target": "es2021",
-		"lib": ["ES2021"],
+		"lib": [
+			"ES2021"
+		],
 		"outDir": "out",
 		"sourceMap": true,
 		"strict": true,
 		"rootDir": "src"
 	},
-	"exclude": ["node_modules", ".vscode-test"]
+	"exclude": [
+		"node_modules",
+		".vscode-test"
+		// "src/pkg"
+	]
 }
