+++ packages/rsx-rosetta/Cargo.toml
@@ -15,7 +15,7 @@ keywords = ["dom", "ui", "gui", "react"]
 [dependencies]
 dioxus-autofmt = { workspace = true }
 dioxus-rsx = { workspace = true }
-html_parser = "0.6.3"
+html_parser.workspace = true
 proc-macro2 = "1.0.49"
 quote = "1.0.23"
 syn = { version = "1.0.107", features = ["full"] }
